refactor: fix PR review issues - code quality, error handling, deduplication
Critical fixes:
- replace unwrap() with proper error handling in backup.rs
- add cross-platform home dir detection (Windows/Unix)
- add mutex for backup history race condition prevention

Code deduplication:
- create shared file_utils module for common operations
- extract hash calculation, checksum verification, file collection
- share utilities between backup and file_copy modules

UX improvements:
- replace alert()/confirm() with console logging
- reduce BackupQueue polling from 5s to 30s
- rely on event listeners for real-time updates

Logging:
- add log crate dependency
- replace eprintln with log::error!/log::warn!

Other:
- remove unused BackupDestination struct
- remove unnecessary clones
- add type documentation for Rust/TS enum mapping
