+++ src-tauri/Cargo.toml
@@ -32,4 +32,5 @@ libc = "0.2"
 sha2 = "0.10"
 chrono = "0.4"
 lazy_static = "1.5"
+log = "0.4"
