+++ src-tauri/src/modules/mod.rs
@@ -1,5 +1,6 @@
 pub mod backup;
 pub mod file_copy;
+pub mod file_utils;
 pub mod import_history;
 pub mod project;
 pub mod sd_card;
